Created new NavToggle component that contains menu button and its labels. Next step: refactoring to use reducer & context instead of state.
