+++ src/components/NavToggle/index.js
@@ -1,0 +1,4 @@
+import s from "./styles.scss";
+
+export { s };
+export { default } from "./NavToggle.jsx";
